+++ src/tools.json
@@ -45,15 +45,6 @@
           "hasExplain": true
         },
         {
-          "id": "Morse",
-          "name": "Morseova abeceda",
-          "desc": "Převod textu na Morseovu abecedu a syntéza zvuku.",
-          "difficulty": "🟢",
-          "path": "crypto/Morse_code.html",
-          "tags": ["šifra", "morse", "zvuk"],
-          "hasExplain": true
-        },
-        {
           "id": "transposition",
           "name": "Transpoziční šifra",
           "desc": "Šifra založená na přeskupení písmen podle klíčového slova.",
@@ -61,6 +52,14 @@
           "path": "crypto/transposition.html",
           "tags": ["šifra", "transpozice", "tabulka"],
           "hasExplain": true
+        },
+        {
+          "id": "affine",
+          "name": "Afinní šifra",
+          "desc": "Lineární šifra kombinující násobení a posun písmen.",
+          "difficulty": "🟡",
+          "path": "crypto/affine_cipher.html",
+          "tags": ["šifra", "afinní", "substituce"]
         }
       ]
     },
@@ -86,6 +85,15 @@
       "chip": "🧮 Kódování",
       "color": "#10b981",
       "tools": [
+        {
+          "id": "Morse",
+          "name": "Morseova abeceda",
+          "desc": "Převod textu na Morseovu abecedu a syntéza zvuku.",
+          "difficulty": "🟢",
+          "path": "encoding/morse_code.html",
+          "tags": ["šifra", "morse", "zvuk"],
+          "hasExplain": true
+        },
         {
           "id": "ascii",
           "name": "ASCII převodník",
